Move tile specific methods to the tile layer renderer
